\name{CancerMenopause}
\alias{CancerMenopause}

\title{
Swedish breast cancer.
}

\description{

Data on incidence of breast cancer among women in Sweden.  Goal of the
study was to investigate whether the incidence increases with the onset
of menopause.

Included here with the permission of Prof. Yudi Pawitan, Karolinska
Institutet, Stockholm.

}

