\name{qe-Series Wrappers}
\alias{DoubleD}

\title{Double Descent Phenomenon}

\description{

Belkin and others have shown that some machine learning algorithms
exhibit surprising behavior when in overfitting settings.  The
\code{doubleD} facilitates easy exploration of this phenomenon.

}


\author{
Norm Matloff
}

