\name{fiftyksongs}
\alias{fiftyksongs}

\title{
Subset of the Million Song Dataset.
}

\description{

Random subset of 50000 songs from the UCI ML Repository dataset.
See 

Thierry Bertin-Mahieux, Daniel P.W. Ellis, Brian Whitman, and Paul Lamere.
The Million Song Dataset. In Proceedings of the 12th International Society
for Music Information Retrieval Conference (ISMIR 2011), 2011,
\url{http://millionsongdataset.com/pages/contact-us/}

}

