\name{iranChurn}
\alias{iranChurn}

\title{
Iranian Customer Churn Data
}

\description{

From \url{https://github.com/sharmaroshan/Churn-Modelling-Dataset}.

Character variables and bernoulli variables have been converted to
factors.  The first three cols, e.g. customer ID, have been deleted.

The tenure col is apparently length of time with the firm.

}

