\name{lsa}
\alias{lsa}

\title{
Law School Admissions Data
}

\description{

Law School Admissions dataset from the Law School Admissions Council
(LSAC).  The dataset was originally collected for a study called 'LSAC
National Longitudinal Bar Passage Study' by Linda Wightman in 1998.

Most of the names are self-explanatory, but note that:  The two
decile scores are class standing in the first and third years of law
school, and 'cluster' refers to the reputed quality of the law school.
Two variables of particular interest might be the student's score on the
Law School Admission Test (LSAT) and a logical variable indicating
whether the person passed the bar examination.

Note that the 'age' variable is apparently birth year, e.g. 69 meaning
1969.

}


