\name{mlb}
\alias{mlb}
\alias{mlb1}

\title{
Major Leage Baseball player data set.
}

\description{
Heights, weights, ages etc. of major league baseball players.  A new
variable has been added, consolidating positions into Infielders,
Outfielders, Catchers and Pitchers.

The \code{mlb1} version has only Position, Height, Weight and Age.

Included here with the permission of the UCLA Statistics Department.
}

