\name{ThyroidDisease}
\alias{ThyroidDisease}

\title{
Thyroid Disease
}

\usage{
data(ThyroidDisease)
}

\description{

See OpenML repository,
https://www.openml.org/search?type=data&sort=runs&id=38&status=active.

"Thyroid disease records supplied by the Garavan Institute
and J. Ross Quinlan, New South Wales Institute, Syndney, Australia.
1987."

}

